requirements.txt
Flask
requests
Pillow
deepface
gunicorn
